Various fixes & improvements (#24)
* Update Cargo.lock

* Reduce libc usage where possible

Use std's types instead of libc's ones

* debug_server: Use truncate method

* Refactor code

* instproxy: Use CStr

We need to use CStr to create a string from an allocated string on the
C side

* mobile_sync: Use CStr instead of CString

We need to create a string from a C allocated string, so we use CStr.

* service: Fix incorrect vector allocation

* Derive more traits for public enums

Derive `Debug, Clone, Copy, PartialEq, Eq` for any public enum for a
better API.

* Derive Debug for some public structs

where it can be useful

* readme: Fix markdown warnings by adding blank lines

* mobile_sync: Fix a memory leak

Previously both `mobilesync_anchors` struct and their `CStrings` were
leaking memory. We can fix this by attaching a `mobilesync_anchors` struct
to `MobileSyncAnchor`, so it will automatically drop everything.

* Cargo fmt

* Fix cargo clippy warnings

* Use slice::from_raw_parts instead of vec::from_raw_parts

Since we get a pointer of an array that was allocated in C, we shouldn't
use vec::from_raw_parts, because it will be eventually dropped: freeing
C allocated memory in Rust (and vice versa) causes an UB. So we create a
slice and clone it (todo: we need to free an old C array as well).

+++ src/connection.rs
@@ -14,6 +14,7 @@ pub struct DeviceConnection<'a> {
 
 pub struct SslData {}
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DeviceConnectionType {
     Usbmuxd,
     Network,
@@ -80,7 +81,7 @@ impl DeviceConnection<'_> {
     ///
     /// ***Verified:*** False
     pub fn receive(&self, len: u32, timeout: Option<u32>) -> Result<Vec<u8>, IdeviceError> {
-        let mut buffer = vec![0 as u8; len as usize];
+        let mut buffer = vec![0_u8; len as usize];
         let mut received = 0;
 
         let result = match timeout {
